fix: split tile radial host catalog under game widgets size gate
Extract layout/spoke builders so game_map_tile_radial_host stays within
repo.game_widgets_file_size (300), unblocking Quality on #4570.
