+++ app/lib/features/game/widgets/map_radial/game_map_tile_radial_host.dart
@@ -3,8 +3,6 @@ library;
 
 import 'package:colonizethis_app/features/game/flame/caches/per_player_work_target_selection_cache.dart';
 import 'package:colonizethis_app/features/game/flame/controls/map_tile_hover_readout_copy.dart';
-import 'package:colonizethis_app/features/game/flame/map_state/game_map_area_state_logic.dart';
-import 'package:colonizethis_app/features/game/flame/map_state/province_action_state_calculator.dart';
 import 'package:colonizethis_app/features/game/flame/overlays/province_detail_overlay_host_support.dart';
 import 'package:colonizethis_app/providers/game_service_provider.dart';
 import 'package:colonizethis_app/providers/games_provider.dart';
@@ -20,10 +18,9 @@ import 'tile_context_radial.dart';
 import 'tile_more_actions_dialog.dart';
 import 'tile_radial_catalog.dart';
 import 'tile_radial_emit.dart';
-import 'tile_radial_keys.dart';
+import 'tile_radial_host_catalog.dart';
 import 'tile_radial_layout.dart';
 import 'tile_radial_spoke_view.dart';
-import 'tile_radial_tooltips.dart';
 
 /// Overlay host: secondary map gesture → radial or More dialog.
 class GameMapTileRadialHost extends ConsumerStatefulWidget {
@@ -88,76 +85,17 @@ class GameMapTileRadialHostState extends ConsumerState<GameMapTileRadialHost> {
   }
 
   TileRadialCatalogLayout _layoutFor(String tileKey) {
-    final draftOrders = ref.read(currentOrdersProvider);
-    final mapData = tryGetGameMapData(
-      () => ref.read(gameServiceProvider).getMapData(widget.game.id),
-    );
-    final states = ProvinceActionStateCalculator.compute(
+    return tileRadialHostCatalogLayout(
       game: widget.game,
       humanPlayerId: widget.humanPlayerId,
-      selectedTileKey: tileKey,
+      tileKey: tileKey,
       region: widget.region,
       playerView: widget.playerView,
-      currentOrders: draftOrders,
       workTargetSelectionCache: widget.workTargetSelectionCache,
-      mapData: mapData,
-    );
-    final parsed = tileKey.split('|');
-    final provinceId = parsed.length >= 2
-        ? '${parsed[0]}|${parsed[1]}'
-        : tileKey;
-    final upgradeTown =
-        GameMapAreaStateLogicProvinceActions.provinceUpgradeTownActionState(
-          game: widget.game,
-          humanPlayerId: widget.humanPlayerId,
-          provinceId: provinceId,
-          playerView: widget.playerView,
-          workTargetSelectionCache: widget.workTargetSelectionCache,
-          topology: mapData?.combinedTopology,
-          currentOrders: draftOrders,
-          tileMapByRegion: mapData?.tileMapByRegion,
-        );
-    final upgradeTownOnSelectedTile =
-        upgradeTown.showControl && upgradeTown.townTileKey == tileKey;
-    return rankTileRadialCatalog(
-      visibility: {
-        TileRadialCatalogAction.explore: (
-          showIcon: states.explore.showIcon,
-          enabled: states.explore.enabled,
-        ),
-        TileRadialCatalogAction.prospect: (
-          showIcon: states.prospect.showIcon,
-          enabled: states.prospect.enabled,
-        ),
-        TileRadialCatalogAction.buildImprovement: (
-          showIcon: states.buildImprovement.showIcon,
-          enabled: states.buildImprovement.enabled,
-        ),
-        TileRadialCatalogAction.buildRoad: (
-          showIcon: states.buildRoad.showIcon,
-          enabled: states.buildRoad.enabled,
-        ),
-        TileRadialCatalogAction.purchaseLand: (
-          showIcon: states.purchaseLand.showIcon,
-          enabled: states.purchaseLand.enabled,
-        ),
-        TileRadialCatalogAction.upgradeTown: (
-          showIcon: upgradeTownOnSelectedTile,
-          enabled: upgradeTownOnSelectedTile && upgradeTown.enabled,
-        ),
-        TileRadialCatalogAction.buildPort: (
-          showIcon: states.buildPort.showIcon,
-          enabled: states.buildPort.enabled,
-        ),
-        TileRadialCatalogAction.buildRail: (
-          showIcon: states.buildRail.showIcon,
-          enabled: states.buildRail.enabled,
-        ),
-        TileRadialCatalogAction.buildFort: (
-          showIcon: states.buildFort.showIcon,
-          enabled: states.buildFort.enabled,
-        ),
-      },
+      draftOrders: ref.read(currentOrdersProvider),
+      mapData: tryGetGameMapData(
+        () => ref.read(gameServiceProvider).getMapData(widget.game.id),
+      ),
     );
   }
 
@@ -166,79 +104,21 @@ class GameMapTileRadialHostState extends ConsumerState<GameMapTileRadialHost> {
     String tileKey,
     List<TileRadialSpoke> spokes,
   ) {
-    final l10n = appL10n(context);
-    final draftOrders = ref.read(currentOrdersProvider);
-    final parsed = tileKey.split('|');
-    final provinceId = parsed.length >= 2
-        ? '${parsed[0]}|${parsed[1]}'
-        : tileKey;
-    final mapData = tryGetGameMapData(
-      () => ref.read(gameServiceProvider).getMapData(widget.game.id),
-    );
-    final states = ProvinceActionStateCalculator.compute(
+    return tileRadialHostSpokeViews(
+      context: context,
+      l10n: appL10n(context),
       game: widget.game,
       humanPlayerId: widget.humanPlayerId,
-      selectedTileKey: tileKey,
+      tileKey: tileKey,
       region: widget.region,
       playerView: widget.playerView,
-      currentOrders: draftOrders,
       workTargetSelectionCache: widget.workTargetSelectionCache,
-      mapData: mapData,
+      draftOrders: ref.read(currentOrdersProvider),
+      mapData: tryGetGameMapData(
+        () => ref.read(gameServiceProvider).getMapData(widget.game.id),
+      ),
+      spokes: spokes,
     );
-    final upgradeTown =
-        GameMapAreaStateLogicProvinceActions.provinceUpgradeTownActionState(
-          game: widget.game,
-          humanPlayerId: widget.humanPlayerId,
-          provinceId: provinceId,
-          playerView: widget.playerView,
-          workTargetSelectionCache: widget.workTargetSelectionCache,
-          topology: mapData?.combinedTopology,
-          currentOrders: draftOrders,
-          tileMapByRegion: mapData?.tileMapByRegion,
-        );
-    bool hasMatchingUnits(TileRadialCatalogAction action) {
-      switch (action) {
-        case TileRadialCatalogAction.explore:
-          return states.explore.hasMatchingUnits;
-        case TileRadialCatalogAction.prospect:
-          return states.prospect.hasMatchingUnits;
-        case TileRadialCatalogAction.buildImprovement:
-          return states.buildImprovement.hasMatchingUnits;
-        case TileRadialCatalogAction.buildRoad:
-          return states.buildRoad.hasMatchingUnits;
-        case TileRadialCatalogAction.purchaseLand:
-          return states.purchaseLand.hasMatchingUnits;
-        case TileRadialCatalogAction.upgradeTown:
-          return upgradeTown.hasBuilderUnits;
-        case TileRadialCatalogAction.buildPort:
-          return states.buildPort.hasMatchingUnits;
-        case TileRadialCatalogAction.buildRail:
-          return states.buildRail.hasMatchingUnits;
-        case TileRadialCatalogAction.buildFort:
-          return states.buildFort.hasMatchingUnits;
-      }
-    }
-
-    return [
-      for (final spoke in spokes)
-        TileRadialSpokeView(
-          action: spoke.action,
-          enabled: spoke.enabled,
-          label: tileRadialActionLabel(l10n, spoke.action),
-          tooltip: tileRadialActionTooltip(
-            context: context,
-            l10n: l10n,
-            action: spoke.action,
-            game: widget.game,
-            humanPlayerId: widget.humanPlayerId,
-            tileKey: tileKey,
-            provinceId: provinceId,
-            currentOrders: draftOrders,
-            enabled: spoke.enabled,
-            hasMatchingUnits: hasMatchingUnits(spoke.action),
-          ),
-        ),
-    ];
   }
 
   void _commit(TileRadialCatalogAction action) {
